+++ scripts/IDE/ide.gd
@@ -38,6 +38,9 @@ func _ready() -> void:
 	vfs = VFS.new()
 	vfs.set_name("ide_custom")
 	
+	get_tree().get_root().size_changed.connect(resize)
+	Globals.IDE = self
+
 	Globals.screen_resized.connect(resize)
 	sidebar_split.dragged.connect(sidebar_to_editor_sync)
 	editor_split.dragged.connect(editor_to_sidebar_sync)